requirements.txt
python-dotenv>=1.1.1
sendgrid>=6.12.4
requests>=2.32.5
flask>=3.0.0
flask-cors>=4.0.0
psutil>=5.9.0
